+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Upcoming
+* Keego
+  * Mining Momentum now requires 24 stacks to be fully stacked instead of 32 (~~+10%~~ -> +15% mining speed per stack)
+  * Increased minimum mining effect time (0.75 -> 1 seconds)
 * Detector Rails no longer need redstone
 
 ## 1.9.6